refactor: change the architecture to allow different search backends
Right now this solves no issue, but I want to experiment with adding
more search backends in the future. For this to be easier, it's probably
good to have an idea of what a "backend" actually is 🙂

I think they can help solve performance issues in massive
repositories. For example, I found out that "git grep" is much faster
than ripgrep in some cases (5s vs 40s).

Related to https://github.com/mikavilpas/blink-ripgrep.nvim/issues/110
