+++ lua/blink-ripgrep/backends/ripgrep/ripgrep_parser.lua
